refactor(CustomError): move CustomError to types file
