+++ src/types/types.d.ts
@@ -92,3 +92,8 @@ declare global {
     };
   }
 }
+
+// Interface to add Error code builtin js error
+interface CustomError extends Error {
+  status?: number;
+}
